refactor: openshift-model-console generated from OpenAPI schemas
