+++ kubernetes-model-generator/openshift-model-console/pom.xml
@@ -60,12 +60,8 @@
             <configuration >
               <settings combine.self="append">
                 <schemas>
-                  <schema>${openapi.schema.openshift-latest}</schema>
+                  <schema>${openapi.schema.openshift-generated}</schema>
                 </schemas>
-                <packageMappings combine.self="append">
-                  <io.k8s.api>io.fabric8.kubernetes.api.model</io.k8s.api>
-                  <io.openshift.console>io.fabric8.openshift.api.model.console</io.openshift.console>
-                </packageMappings>
                 <includeGenerationRegexes>
                   <includeGenerationRegex>^io\.openshift\.console\..*$</includeGenerationRegex>
                 </includeGenerationRegexes>
